LayoutTests/platform/ios-simulator/editing/selection/move-by-word-visually-multi-space-expected.txt

CONSOLE MESSAGE: line 16: TypeError: null is not an object (evaluating 'positions[i].node.tagName')
abc def hij opq
abc def hij opq
abc אבצ def
abc def אבצ דדפ hij opq
abc def hij אבצ דעפ היח opq rst uvw
אבצ דעפ חיח ופק
אבצ abc דעפ
אבצ דעפ abc def חיח ופק
אבצ דעפ חיח abc def jih ופק ופק רסת
אבצ דעפ היח ופק
אבצ דעפ היח abc def hij
abc def hij אבצ דעפ היח opq rst uvw
abc def hij opq
abc def hij opq
abc אבצ def
abc def אבצ דדפ hij opq
abc def hij אבצ דעפ היח opq rst uvw
אבצ דעפ חיח ופק
אבצ abc דעפ
אבצ דעפ abc def חיח ופק
אבצ דעפ חיח abc def jih ופק ופק רסת
abc def    hij opq
    abc    def    hij    opq    
    abc    אבצ    def    
    abc    def    אבצ    דדפ    hij    opq    
    abc    def    hij    אבצ    דעפ    היח    opq    rst    uvw    
    אבצ    דעפ    חיח    ופק    
    אבצ    abc    דעפ     
    אבצ    דעפ    abc   def   חיח    ופק    
    אבצ    דעפ    חיח    abc   def   jih   ופק  ופק    רסת     
אבצ דעפ    היח ופק
אבצ דעפ היח    abc def hij
abc def hij    אבצ דעפ היח    opq rst uvw
abc def    hij opq
    abc    def    hij    opq    
    abc    אבצ    def    
    abc    def    אבצ    דדפ    hij    opq    
    abc    def    hij    אבצ    דעפ    היח    opq    rst    uvw    
    אבצ    דעפ    חיח    ופק    
    אבצ    abc    דעפ     
    אבצ    דעפ    abc   def   חיח    ופק    
    אבצ    דעפ    חיח    abc   def   jih   ופק  ופק    רסת     
人一氧喝大    笑抬的     abc def

======== Move By Word ====
Test 1, LTR:
Move right by one word
"abc def    hij opq"[0, 4, 8, 15, 18]
Move left by one word
"abc def    hij opq"[18, 15, 8, 4, 0]
Test 2, LTR:
Move right by one word
"    abc    def    hij    opq    "[4, 8, 15, 22, 28]
Move left by one word
"    abc    def    hij    opq    "[28, 22, 15, 8, 4]
Test 3, LTR:
Move right by one word
"    abc    ABW    def    "[4, 8, 15, 21]
Move left by one word
"    abc    ABW    def    "[21, 15, 8, 4]
Test 4, LTR:
Move right by one word
"    abc    def    ABW    DDU    hij    opq    "[4, 8, 15, 21, 29, 36, 42]
Move left by one word
"    abc    def    ABW    DDU    hij    opq    "[42, 36, 29, 21, 15, 8, 4]
Test 5, LTR:
Move right by one word
"    abc    def    hij    ABW    DSU    EJH    opq    rst    uvw    "[4, 8, 15, 22, 35, 28, 43, 50, 57, 63]
Move left by one word
"    abc    def    hij    ABW    DSU    EJH    opq    rst    uvw    "[63, 57, 50, 43, 28, 35, 22, 15, 8, 4]
Test 6, LTR:
Move right by one word
"    ABW    DSU    HJH    FUX    "[4, 21, 14, 7, 28]
Move left by one word
"    ABW    DSU    HJH    FUX    "[28, 7, 14, 21, 4]
Test 7, LTR:
Move right by one word
"    ABW    abc    DSU     "[4, 8, 15, 21]
Move left by one word
"    ABW    abc    DSU     "[21, 15, 8, 4]
Test 8, LTR:
Move right by one word
"    ABW    DSU    abc   def   HJH    FUX    "[4, 7, 15, 22, 28, 33, 40]
Move left by one word
"    ABW    DSU    abc   def   HJH    FUX    "[40, 33, 28, 22, 15, 7, 4]
Test 9, LTR:
Move right by one word
"    ABW    DSU    HJH    abc   def   jih   FUX  FUX    YR[     "[4, 14, 7, 22, 29, 35, 41, 51, 46, 58]
Move left by one word
"    ABW    DSU    HJH    abc   def   jih   FUX  FUX    YR[     "[58, 46, 51, 41, 35, 29, 22, 7, 14, 4]
Test 10, LTR:
Move right by one word
"ABW DSU    EJH FUX"[0, 14, 7, 3, 18]
Move left by one word
"ABW DSU    EJH FUX"[18, 3, 7, 14, 0]
Test 11, LTR:
Move right by one word
"ABW DSU EJH    abc def hij"[0, 7, 3, 12, 19, 23, 26]
Move left by one word
"ABW DSU EJH    abc def hij"[26, 23, 19, 12, 3, 7, 0]
Test 12, LTR:
Move right by one word
"abc def hij    ABW DSU EJH    opq rst uvw"[0, 4, 8, 12, 22, 18, 27, 34, 38, 41]
Move left by one word
"abc def hij    ABW DSU EJH    opq rst uvw"[41, 38, 34, 27, 18, 22, 12, 8, 4, 0]
Test 13, RTL:
Move left by one word
"abc def    hij opq"[0, 14, 7, 3, 18]
Move right by one word
"abc def    hij opq"[18, 3, 7, 14, 0]
Test 14, RTL:
Move left by one word
"    abc    def    hij    opq    "[4, 21, 14, 7, 28]
Move right by one word
"    abc    def    hij    opq    "[28, 7, 14, 21, 4]
Test 15, RTL:
Move left by one word
"    abc    ABW    def    "[4, 8, 15, 21]
Move right by one word
"    abc    ABW    def    "[21, 15, 8, 4]
Test 16, RTL:
Move left by one word
"    abc    def    ABW    DDU    hij    opq    "[4, 7, 15, 22, 29, 35, 42]
Move right by one word
"    abc    def    ABW    DDU    hij    opq    "[42, 35, 29, 22, 15, 7, 4]
Test 17, RTL:
Move left by one word
"    abc    def    hij    ABW    DSU    EJH    opq    rst    uvw    "[4, 14, 7, 22, 29, 36, 43, 56, 49, 63]
Move right by one word
"    abc    def    hij    ABW    DSU    EJH    opq    rst    uvw    "[63, 49, 56, 43, 36, 29, 22, 7, 14, 4]
Test 18, RTL:
Move left by one word
"    ABW    DSU    HJH    FUX    "[4, 8, 15, 22, 28]
Move right by one word
"    ABW    DSU    HJH    FUX    "[28, 22, 15, 8, 4]
Test 19, RTL:
Move left by one word
"    ABW    abc    DSU     "[4, 8, 15, 21]
Move right by one word
"    ABW    abc    DSU     "[21, 15, 8, 4]
Test 20, RTL:
Move left by one word
"    ABW    DSU    abc   def   HJH    FUX    "[4, 8, 15, 21, 28, 34, 40]
Move right by one word
"    ABW    DSU    abc   def   HJH    FUX    "[40, 34, 28, 21, 15, 8, 4]
Test 21, RTL:
Move left by one word
"    ABW    DSU    HJH    abc   def   jih   FUX  FUX    YR[     "[4, 8, 15, 22, 34, 28, 41, 47, 52, 58]
Move right by one word
"    ABW    DSU    HJH    abc   def   jih   FUX  FUX    YR[     "[58, 52, 47, 41, 28, 34, 22, 15, 8, 4]
Test 22, LTR:
Move right by one word
"abc def    hij opq"[0, 4, 11, 15, 18]
Move left by one word
"abc def    hij opq"[18, 15, 11, 4, 0]
Test 23, LTR:
Move right by one word
"    abc    def    hij    opq    "[0, 4, 11, 18, 25, 32]
Move left by one word
"    abc    def    hij    opq    "[32, 25, 18, 11, 4, 0]
Test 24, LTR:
Move right by one word
"    abc    ABW    def    "[0, 4, 11, 18, 25]
Move left by one word
"    abc    ABW    def    "[25, 18, 11, 4, 0]
Test 25, LTR:
Move right by one word
"    abc    def    ABW    DDU    hij    opq    "[0, 4, 11, 18, 21, 32, 39, 46]
Move left by one word
"    abc    def    ABW    DDU    hij    opq    "[46, 39, 32, 21, 18, 11, 4, 0]
Test 26, LTR:
Move right by one word
